third_party/WebKit/LayoutTests/fast/dom/HTMLImageElement/image-picture-1x-expected.txt

PASS successfullyParsed is true

TEST COMPLETE
PASS document.getElementById("canary").clientWidth is 75
PASS document.getElementById("canary").currentSrc is "resources/blue_rect.jpg?0"
PASS document.getElementById("simple").clientWidth is 800
PASS document.getElementById("simple").currentSrc is "../../hidpi/resources/image-set-2x.png?2"
PASS document.getElementById("multiple").clientWidth is 800
PASS document.getElementById("multiple").currentSrc is "../../hidpi/resources/image-set-2x.png?2"
PASS document.getElementById("other_children").clientWidth is 800
PASS document.getElementById("other_children").currentSrc is "../../hidpi/resources/image-set-2x.png?2"
PASS document.getElementById("img_before_source").clientWidth is 75
PASS document.getElementById("img_before_source").currentSrc is "resources/blue_rect.jpg?1"
PASS document.getElementById("unknown_type").clientWidth is 1600
PASS document.getElementById("unknown_type").currentSrc is "resources/image-set-4x.png?3"
PASS document.getElementById("empty_srcset").clientWidth is 1600
PASS document.getElementById("empty_srcset").currentSrc is "resources/image-set-4x.png?3"
PASS document.getElementById("no_srcset").clientWidth is 1600
PASS document.getElementById("no_srcset").currentSrc is "resources/image-set-4x.png?3"
   
some text
     
